Refatorando algumas partes do codigo
