+++ App/controllers/HomeController.php
@@ -11,21 +11,7 @@ class HomeController extends Controller {
 
     public function editar($id) {
         $usuario = User::getById($id);
-        $this->view('home/editar_usuario.php',[
-            'usuario' => $usuario
-        ]);
+        $this->view('home/editar_usuario.php',['usuario' => $usuario]);
     }
 
-    public function atualizar() {
-        require_once '../App/models/User.php';
-        $id = $_POST['id'];
-        $nome = $_POST['nome'];
-        $cpf = $_POST['cpf'];
-        $plano = $_POST['plano'];
-        $status = $_POST['status'];
-
-        User::update($id, $nome, $cpf, $plano, $status);
-        header('Location: index.php');
-        exit;
-    }
 }
